Add time to output for "who's on" query with no one on
To do this, and in preparation for the next commit, use the originally parsed
time rather than reparsing based on the schedule api response.

I'm not sure what we gained by re-parsing the time from the response, and I
think one of these tests was wrong. Eg 'should work on happy path' did a query
for 3pm Karl's time and then expected the response to describe 10pm UTC.
Confusion #1 is that Karl's timezone seems to be UTC according to the mocked
local db, and UTC-0700 according to the mocked schedule api response. There also
may have been an idea that the schedule api would return a wider range of times
describing the whole shift, but in my testing the api returns the 1m range that
the code requests.

I'm not 100% sure that I'm not missing anything though so if you can think of
why we would want the old behavior, let me know.